deployment/requirements.txt
# CDK dependencies
aws-cdk-lib>=2.166.0
aws-cdk.aws-lambda-python-alpha==2.166.0-alpha.0
constructs>=10.0.0

# AWS SDK and utilities
boto3>=1.37.4
botocore>=1.37.4

# Development dependencies
pytest>=7.0.0
black>=22.0.0
pylint>=2.15.0
requests_aws4auth>=1.3.1
opensearch-py>=2.8.0

prettytable>=3.12.0
# streamlit>=1.40.0
